feat: 파일 위치 변경(자원별로 묶음)
