+++ src/main/java/com/multi/mlpenterpriseapprovalsystem/reservation/meetingroom/domain/MeetingRoomReservation.java
@@ -1,8 +1,7 @@
-package com.multi.mlpenterpriseapprovalsystem.reservation.reserve.domain;
+package com.multi.mlpenterpriseapprovalsystem.reservation.meetingroom.domain;
 
 import com.multi.mlpenterpriseapprovalsystem.company.domain.Company;
 import com.multi.mlpenterpriseapprovalsystem.employee.domain.Employee;
-import com.multi.mlpenterpriseapprovalsystem.reservation.register.domain.MeetingRoom;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
