+++ src/main/java/com/multi/mlpenterpriseapprovalsystem/reservation/meetingroom/dto/ReqMeetingRoomDto.java
@@ -1,4 +1,4 @@
-package com.multi.mlpenterpriseapprovalsystem.reservation.dto;
+package com.multi.mlpenterpriseapprovalsystem.reservation.meetingroom.dto;
 
 import lombok.*;
