+++ src/main/java/com/multi/mlpenterpriseapprovalsystem/reservation/share_equipment/domain/SharedEquipment.java
@@ -1,4 +1,4 @@
-package com.multi.mlpenterpriseapprovalsystem.reservation.register.domain;
+package com.multi.mlpenterpriseapprovalsystem.reservation.share_equipment.domain;
 
 import com.multi.mlpenterpriseapprovalsystem.company.domain.Company;
 import jakarta.persistence.*;
@@ -10,20 +10,18 @@ import lombok.NoArgsConstructor;
  * Please explain the class!!!
  *
  * @author : 김승기
- * @filename : CorporateCar
+ * @filename : SharedEquipment
  * @since : 2025. 12. 16. 화요일
  */
-@Entity
-@Getter
+@Entity @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "corporate_car")
-public class CorporateCar {
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY) private Long carNo;
+@Table(name = "shared_equipment")
+public class SharedEquipment {
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY) private Long eqNo;
     @ManyToOne(fetch = FetchType.LAZY) @JoinColumn(name = "com_id", referencedColumnName = "com_id") private Company company;
-    private String carName;
-    private String carType;
-    private String plateNo;
-    private Integer cap;
-    private String fuel;
+    private String eqName;
+    private String eqId;
+    private String modelName;
     private String imgUrl;
+    private String loc;
 }
