+++ src/main/java/com/multi/mlpenterpriseapprovalsystem/reservation/share_equipment/domain/SharedEquipmentReservation.java
@@ -1,8 +1,7 @@
-package com.multi.mlpenterpriseapprovalsystem.reservation.reserve.domain;
+package com.multi.mlpenterpriseapprovalsystem.reservation.share_equipment.domain;
 
 import com.multi.mlpenterpriseapprovalsystem.company.domain.Company;
 import com.multi.mlpenterpriseapprovalsystem.employee.domain.Employee;
-import com.multi.mlpenterpriseapprovalsystem.reservation.register.domain.CorporateCar;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -14,17 +13,17 @@ import java.time.LocalDateTime;
  * Please explain the class!!!
  *
  * @author : 김승기
- * @filename : CorporateCarReservation
+ * @filename : SharedEquipmentReservation
  * @since : 2025. 12. 16. 화요일
  */
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "corporate_car_reservation")
-public class CorporateCarReservation {
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY) private Long carResvNo;
+@Table(name = "shared_equipment_reservation")
+public class SharedEquipmentReservation {
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY) private Long eqResvNo;
     @ManyToOne(fetch = FetchType.LAZY) @JoinColumn(name = "com_id", referencedColumnName = "com_id") private Company company;
-    @ManyToOne(fetch = FetchType.LAZY) @JoinColumn(name = "car_no") private CorporateCar corporateCar;
+    @ManyToOne(fetch = FetchType.LAZY) @JoinColumn(name = "eq_no") private SharedEquipment sharedEquipment;
     private LocalDateTime startedAt;
     private LocalDateTime endedAt;
     @ManyToOne(fetch = FetchType.LAZY) @JoinColumn(name = "resv_emp", referencedColumnName = "emp_id") private Employee resvEmp;
